laser_frame and equations Maartje

+++ Data_generation/generate_background.m
@@ -1,4 +1,4 @@
 function frame = generate_background(frame, set)
 
-frame = frame + poissrnd(set.other.av_background,[set.mic.pixels_y,set.mic.pixels_x]);
+frame = frame + int16(poissrnd(set.other.av_background,[set.mic.pixels_y,set.mic.pixels_x]));
 end
